+++ odra-vm/src/odra_vm_host.rs
@@ -198,7 +198,8 @@ impl HostContext for OdraVmHost {
         contract_name: String,
         entry_points_caller: EntryPointsCaller
     ) {
-        panic!("register_contract is not supported for OdraVM");
+        self.vm
+            .register_contract(address, &contract_name, entry_points_caller);
     }
 
     fn contract_env(&self) -> ContractEnv {